+++ build_runner/lib/src/build_plan/build_package.dart
@@ -17,9 +17,6 @@ class BuildPackage {
   /// Paths are platform dependent.
   final String path;
 
-  /// Whether the "package" is actually a workspace.
-  final bool isWorkspace;
-
   /// Whether the package contents should be watched in `watch` and `serve`
   /// modes.
   final bool watch;
@@ -37,7 +34,6 @@ class BuildPackage {
   BuildPackage({
     required this.name,
     required String path,
-    this.isWorkspace = false,
     this.watch = false,
     this.isOutput = false,
     this.languageVersion,
@@ -56,7 +52,6 @@ class BuildPackage {
   }) => BuildPackage(
     name: name,
     path: '/$name',
-    isWorkspace: isWorkspace,
     watch: watch,
     isOutput: isOutput,
     dependencies: dependencies,
@@ -68,7 +63,6 @@ class BuildPackage {
     return other is BuildPackage &&
         name == other.name &&
         path == other.path &&
-        isWorkspace == other.isWorkspace &&
         watch == other.watch &&
         isOutput == other.isOutput &&
         languageVersion == other.languageVersion &&
@@ -84,7 +78,6 @@ class BuildPackage {
 BuildPackage(
   name: $name,
   path: $path,
-  isWorkspace: $isWorkspace,
   watch: $watch,
   isOutput: $isOutput,
   languageVersion: $languageVersion,